requirements.txt
Scrapy>=2.11
scrapy-playwright>=0.0.35
pandas>=2.2
SQLAlchemy>=2.0
psycopg2-binary>=2.9
beautifulsoup4>=4.12
requests>=2.31
PyYAML>=6.0
python-dateutil>=2.8
